Don't clear identities when making a new terminal
We had previously told keychain to clear all saved identities, so that
each new login would require a password for the keychain. In practice,
it didn't work out that way. Whenever we made a new xterm, we'd get
prompted for our password. This change removes the "--clear" option,
so that we don't have to input our password whenever we open a new
xterm.