tmux: Use F5 to scroll back to previous prompt
When working with commands that emit lots of text to the console, it's
convenient to jump back up to the most recent previous command prompt.
This lets you read the output of the previously run command from the
beginning, avoiding having to manually search for the start by pressing
page up a lot.