kitty: Don't check for automatic updates
We use the native package manager to keep our software up to date.
Package maintainers are supposed to disable update checking at
build-time, but sometimes they forget or don't know about this
feature.

From kitty docs:
    Update checking is only done by the official binary builds. Distro
    packages or source builds do not do update checking.

But, for package builds not to do update checking, the package
maintainer must remember to build kitty with the option
"--update_check_interval=0".
    python3 setup.py linux-package --update-check-interval=0

Add this option to our config as a backstop in case the package
maintainer misses disabling this feature.