nixos: Add perlPackages.URI
We add a new section to environment.systemPackages to specify which
packages we want our global perl environment to have. Without this, even
if both perl and perlPackages.URI are explicitly installed in the user
environment, perl fails to find the URI module.

    Can't locate URI.pm in @INC (you may need to install the URI module)

Installing the perl module globally was decided as a good way forward
because it did not require editing exising perl scripts with a
"nix-shell" shebang, nor always running perl scripts with a `nix-shell
-p perl perlPackages.URI --run ./somescript.pl` invocation.