urxvt: Configure cutchars
When double-clicking on a word that's next to a tmux window pane
(delimited by a '│' with our current configuration), the pane delimiter
is selected in addition to the word. We want to exclude the delimiting
character from word selection, so we specify a non-default cutchars
which includes separating words on our delimiter ('│').